[refactor]: Re-organise `sumeragi`.
